refactor(web): colocate page, routes. adjust naming convention
