+++ apps/web/app/constants/constants.ts
@@ -43,4 +43,4 @@ export const STACKS_BLOCKS_PER_DAY = 144;
 
 export const STACKING_TRACKER_API_URL = 'https://api.stacking-tracker.com';
 
-export const DASH = '—';
+export const EM_DASH = '—';
