+++ apps/web/app/pages/stacking/pooled/pooled-stacking.route.tsx
@@ -4,7 +4,7 @@ import { StackingClientProvider } from '~/features/stacking/providers/stacking-c
 import { poolSlugSchema } from '~/features/stacking/start-pooled-stacking/utils/stacking-pool-types';
 import { PooledStacking } from '~/pages/stacking/pooled/pooled-stacking';
 
-import { Route } from './+types/pooled-stacking.page';
+import { Route } from './+types/pooled-stacking.route';
 
 export function loader({ params }: Route.LoaderArgs) {
   const { success, data: poolSlug } = poolSlugSchema.safeParse(params.slug);
